Rewrite About component to use hooks
